simple_adaptive_research_demo.py
"""
Simple Adaptive Research System Demo

This script demonstrates the implementation of the AI-Driven Adaptive Research System (ADR)
components as outlined in the upgrade plan, without complex inter-module dependencies.
"""

import asyncio
import json
from typing import Dict, Any, List, Optional
from datetime import datetime
import logging
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Import the implemented components directly
from ai_agent.src.agents.base.osint_agent import AgentConfig
from ai_agent.src.agents.intelligence_planning.ai_planner_agent import AIPlannerAgent, ProjectPlan
from ai_agent.src.agents.specialized.objective_definition_agent import ObjectiveDefinitionAgent, ObjectiveDefinition
from ai_agent.src.agents.specialized.strategy_formulation_agent import StrategyFormulationAgent, ResearchStrategy
from ai_agent.src.agents.specialized.data_collection_agent import DataCollectionAgent, CollectionPlan


class SimpleAdaptiveResearchSystem:
    """
    Simple Adaptive Research System - Demonstrates the core components
    of the AI-Driven Adaptive Research System without complex dependencies.
    """
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        
        # Initialize components
        self.planner_agent = AIPlannerAgent(
            config=AgentConfig(
                role="AI Planner Agent",
                description="Specialized agent for creating intelligent research project plans",
                llm_model="gpt-4-turbo"
            )
        )
        
        self.objective_agent = ObjectiveDefinitionAgent(
            config=AgentConfig(
                role="Objective Definition Agent",
                description="Specialized agent for defining clear research objectives",
                llm_model="gpt-4-turbo"
            )
        )
        
        self.strategy_agent = StrategyFormulationAgent(
            config=AgentConfig(
                role="Strategy Formulation Agent",
                description="Specialized agent for creating research strategies",
                llm_model="gpt-4-turbo"
            )
        )
        
        self.collection_agent = DataCollectionAgent(
            config=AgentConfig(
                role="Data Collection Agent",
                description="Specialized agent for intelligent data gathering",
                llm_model="gpt-4-turbo"
            )
        )
    
    async def run_complete_research_cycle(self, research_topic: str, initial_requirements: List[str]):
        """
        Run a complete research cycle demonstrating all implemented components
        """
        print(f"🚀 Starting research cycle for: {research_topic}")
        print("="*60)
        
        # Step 1: Project Planning
        print("📋 Step 1: Project Planning")
        project_plan = await self.planner_agent.create_project(
            research_topic=research_topic,
            initial_objectives=initial_requirements,
            project_scope="Comprehensive analysis with adaptive capabilities"
        )
        print(f"✅ Project Plan: {project_plan.project_name}")
        print(f"🎯 Objectives: {len(project_plan.research_objectives)} defined")
        print()
        
        # Step 2: Objective Definition
        print("🎯 Step 2: Objective Definition")
        objective_def = await self.objective_agent.define_objectives(
            research_topic=research_topic,
            research_scope=project_plan.project_description,
            initial_requirements=project_plan.research_objectives
        )
        print(f"✅ Primary Objective: {objective_def.primary_objective}")
        print(f"✅ Secondary Objectives: {len(objective_def.secondary_objectives)}")
        print(f"✅ Success Criteria: {len(objective_def.success_criteria)} defined")
        print()
        
        # Step 3: Strategy Formulation
        print("🧠 Step 3: Strategy Formulation")
        strategy = await self.strategy_agent.formulate_strategy(
            objectives=[objective_def.primary_objective] + objective_def.secondary_objectives,
            resources={"personnel": 1, "time": "30 days", "tools": ["default"]},
            constraints=["time", "access", "quality"]
        )
        print(f"✅ Strategy: {strategy.strategy_name}")
        print(f"✅ Methodology: {strategy.approach_methodology}")
        print(f"✅ Phases: {len(strategy.phase_timeline)} defined")
        print()
        
        # Step 4: Data Collection Planning
        print("🔍 Step 4: Data Collection Planning")
        collection_plan = await self.collection_agent.plan_collection(
            research_objectives=[objective_def.primary_objective],
            strategy_guidelines=strategy.dict(),
            resource_constraints={"bandwidth": "standard", "tools": ["default"]}
        )
        print(f"✅ Sources to Query: {len(collection_plan.sources_to_query)}")
        print(f"✅ Data Types: {collection_plan.data_types_requested}")
        print(f"✅ Quality Criteria: {len(collection_plan.quality_criteria)} defined")
        print()
        
        # Step 5: Data Collection Execution
        print("📥 Step 5: Data Collection Execution")
        collected_data = await self.collection_agent.collect_data(collection_plan)
        print(f"✅ Collection Status: {collected_data.get('collection_status', 'unknown')}")
        print(f"✅ Sources Accessed: {len(collected_data.get('sources_accessed', []))}")
        print(f"✅ Data Types Collected: {len(collected_data.get('data_types_collected', []))}")
        print()
        
        # Compile and return results
        results = {
            "project_plan": project_plan.dict(),
            "objectives": objective_def.dict(),
            "strategy": strategy.dict(),
            "collection_plan": collection_plan.dict(),
            "collected_data": collected_data,
            "execution_timestamp": datetime.utcnow().isoformat()
        }
        
        return results


async def demonstrate_adaptive_system():
    """
    Demonstrate the implemented Adaptive Research System components
    """
    print("AI-Driven Adaptive Research System (ADR) - Implementation Demonstration")
    print("="*80)
    print()
    
    # Initialize the system
    research_system = SimpleAdaptiveResearchSystem()
    
    # Define a research topic
    research_topic = "Analysis of Renewable Energy Adoption in European Markets"
    requirements = [
        "Assess current renewable energy adoption rates",
        "Identify key market drivers and barriers", 
        "Analyze policy impacts on adoption",
        "Evaluate future growth projections"
    ]
    
    print(f"🔬 Research Topic: {research_topic}")
    print(f"📋 Requirements: {requirements}")
    print()
    
    # Run the complete research cycle
    results = await research_system.run_complete_research_cycle(research_topic, requirements)
    
    print("✅ Research Cycle Completed Successfully!")
    print("="*80)
    print()
    
    # Display key achievements
    project_plan = results['project_plan']
    objectives = results['objectives']
    strategy = results['strategy']
    
    print("🏆 System Achievements:")
    print(f"  • Project Plan: {project_plan['project_name']}")
    print(f"  • Primary Objective: {objectives['primary_objective']}")
    print(f"  • Strategy Formulated: {strategy['strategy_name']}")
    print(f"  • Sources Identified: {len(results['collection_plan']['sources_to_query'])}")
    print(f"  • Data Collection Executed: {results['collected_data']['collection_status']}")
    print()
    
    print("🎯 ADR System Capabilities Implemented:")
    print("  1. ✅ AI Planner Agent - Creates intelligent project plans")
    print("  2. ✅ Objective Definition Agent - Defines clear research objectives") 
    print("  3. ✅ Strategy Formulation Agent - Creates comprehensive research strategies")
    print("  4. ✅ Data Collection Agent - Executes intelligent data gathering")
    print("  5. ✅ Adaptive Architecture - Components work together seamlessly")
    print()
    
    print("🔄 Next Steps (Per Upgrade Plan):")
    print("  • Implement Project Orchestrator for task management")
    print("  • Build Research Loop Engine for iterative processes")
    print("  • Create additional specialized agents")
    print("  • Implement adaptive feedback mechanisms")
    print("  • Add real-time monitoring and adjustment capabilities")
    
    print()
    print("✨ The AI-Driven Adaptive Research System (ADR) architecture is now implemented!")


if __name__ == "__main__":
    # Set up logging
    logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
    
    # Run the demonstration
    asyncio.run(demonstrate_adaptive_system())